Changes for story 689, migration of structs DiscoveredPublication_SecurityWrapper and DiscoveredSubscription_SecurityWrapper. Included movement of types to MessageTypes and method inlines to BaseMessageUtils. Put wrapper methods inside ParameterListConverter namespace. Also removal of all references to old source locations of such.
