+++ dds/DCPS/RTPS/BaseMessageUtils.h
@@ -206,6 +206,71 @@ inline void append_submessage(RTPS::Message& message, const RTPS::DataFragSubmes
   DCPS::push_back(message.submessages, sm);
 }
 
+#ifdef OPENDDS_SECURITY
+inline DDS::Security::ParticipantSecurityAttributesMask
+security_attributes_to_bitmask(const DDS::Security::ParticipantSecurityAttributes& sec_attr)
+{
+  using namespace DDS::Security;
+
+  ParticipantSecurityAttributesMask result = PARTICIPANT_SECURITY_ATTRIBUTES_FLAG_IS_VALID;
+  if (sec_attr.is_rtps_protected) {
+    result |= PARTICIPANT_SECURITY_ATTRIBUTES_FLAG_IS_RTPS_PROTECTED;
+  }
+  if (sec_attr.is_discovery_protected) {
+    result |= PARTICIPANT_SECURITY_ATTRIBUTES_FLAG_IS_DISCOVERY_PROTECTED;
+  }
+  if (sec_attr.is_liveliness_protected) {
+    result |= PARTICIPANT_SECURITY_ATTRIBUTES_FLAG_IS_LIVELINESS_PROTECTED;
+  }
+  return result;
+}
+
+inline DDS::Security::EndpointSecurityAttributesMask
+security_attributes_to_bitmask(const DDS::Security::EndpointSecurityAttributes& sec_attr)
+{
+  using namespace DDS::Security;
+
+  EndpointSecurityAttributesMask result = ENDPOINT_SECURITY_ATTRIBUTES_FLAG_IS_VALID;
+
+  if (sec_attr.base.is_read_protected)
+    result |= ENDPOINT_SECURITY_ATTRIBUTES_FLAG_IS_READ_PROTECTED;
+
+  if (sec_attr.base.is_write_protected)
+    result |= ENDPOINT_SECURITY_ATTRIBUTES_FLAG_IS_WRITE_PROTECTED;
+
+  if (sec_attr.base.is_discovery_protected)
+    result |= ENDPOINT_SECURITY_ATTRIBUTES_FLAG_IS_DISCOVERY_PROTECTED;
+
+  if (sec_attr.base.is_liveliness_protected)
+    result |= ENDPOINT_SECURITY_ATTRIBUTES_FLAG_IS_LIVELINESS_PROTECTED;
+
+  if (sec_attr.is_submessage_protected)
+    result |= ENDPOINT_SECURITY_ATTRIBUTES_FLAG_IS_SUBMESSAGE_PROTECTED;
+
+  if (sec_attr.is_payload_protected)
+    result |= ENDPOINT_SECURITY_ATTRIBUTES_FLAG_IS_PAYLOAD_PROTECTED;
+
+  if (sec_attr.is_key_protected)
+    result |= ENDPOINT_SECURITY_ATTRIBUTES_FLAG_IS_KEY_PROTECTED;
+
+  return result;
+}
+
+inline DDS::OctetSeq
+handle_to_octets(DDS::Security::NativeCryptoHandle handle)
+{
+  DDS::OctetSeq handleOctets(sizeof handle);
+  handleOctets.length(handleOctets.maximum());
+  unsigned char* rawHandleOctets = handleOctets.get_buffer();
+  unsigned int handleTmp = handle;
+  for (unsigned int j = sizeof handle; j > 0; --j) {
+    rawHandleOctets[j - 1] = handleTmp & 0xff;
+    handleTmp >>= 8;
+  }
+  return handleOctets;
+}
+#endif
+
 } // namespace RTPS
 } // namespace OpenDDS
